requirements.txt
aiofiles==0.8.0
aioredis==2.0.1
amqp==5.1.1
anyio==3.6.1
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
fastapi==0.82.0
ffmpeg-python==0.2.0
future==0.18.2
h11==0.13.0
idna==3.3
kombu==5.2.4
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
prompt-toolkit==3.0.31
pydantic==1.10.1
pyparsing==3.0.9
python-multipart==0.0.5
pytz==2022.2.1
redis==4.3.4
six==1.16.0
sniffio==1.3.0
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.3.0
ujson==5.4.0
uvicorn==0.18.3
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
